Refactor the applying-interactions tutorial.
